pyproject.toml
[tool.poetry]
name = "babyhelm"
version = "0.1.1"
description = "Babyhelm project backend"
authors = ["i.shamsutdinov", "n.bagrov", "s.lekhtin", "e.solodovnikov", "e.salnikov"]
license = "LICENSE"
readme = "README.md"

[tool.poetry.scripts]
babyhelm-server = "babyhelm.app:main"

[tool.poetry.extras]
lint = ["black", "flake8", "isort"]

[tool.poetry.group.lint.dependencies]
black = "^23.12.1"
flake8 = "^6.1.0"
flake8-comprehensions = "^3.14.0"
flake8-isort = "^6.1.1"
flake8-pytest-style = "^1.7.2"
isort = "^5.13.2"
pep8-naming = "^0.13.3"
pre-commit = "^3.6.0"

[tool.isort]
profile = "black"
known_first_party = ["babyhelm"]
skip = [".gitignore", ".dockerignore"]
extend_skip = [".md", ".json", ".rst"]

[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
aiosqlite = "^0.19.0"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
dependency-injector = "^4.41.0"
jinja2 = "^3.1.4"
httpx = "^0.27.0"
pytest-cov = "^5.0.0"
sqlalchemy = "^2.0.30"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
greenlet = "^3.0.3"
kubernetes = "^30.1.0"
prometheus-fastapi-instrumentator = "^7.0.0"
sentry-sdk = "^2.5.1"
tox = "^4.15.1"

pre-commit = "^3.7.1"
pyjwt = "^2.8.0"
[tool.poetry.group.dev.dependencies]
ruff = "^0.4.8"
pytest = "^8.2.2"
taskipy = "^1.12.2"

[tool.taskipy.tasks]
babyhelm-dev = { cmd = "babyhelm-server --config ./local/config.dev.yaml --reload --log-level=debug", help = "Start BabyHelm dev server" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
omit = [
    "tests/*",
]
